Removing Conditional Logic
Removing Conditional Logic for EnsureRedirectionsList. This method uses the csom ensure() method, which is intended to find a list and create it if it doesn't exist. However, it requires elevated permissions to run. As a result, users with Read permissions to the Redirection List cannot use this redirect functionality. 

I removed the if statement wrapping this method, and added a catch for logging the error. Users with permission will be able to execute this. Read only users it will silently fail with a 403, and then continue processing under the assumption that the list is already created.